requirements.txt
fastapi
uvicorn
openai==0.28
gspread
google-auth
google-auth-oauthlib
requests
python-dotenv
